+++ firebase-auth/src/main/java/com/androidhuman/rxfirebase2/auth/RxFirebaseAuth.java
@@ -14,6 +14,7 @@ import io.reactivex.Completable;
 import io.reactivex.Maybe;
 import io.reactivex.Observable;
 import io.reactivex.Single;
+import io.reactivex.functions.Function;
 
 
 public final class RxFirebaseAuth {
@@ -62,30 +63,82 @@ public final class RxFirebaseAuth {
 
     @CheckResult
     @NonNull
-    public static Single<FirebaseUser> signInAnonymously(
+    public static Single<AuthResult> signInAnonymouslyAuthResult(
             @NonNull FirebaseAuth instance) {
         return new SignInAnonymouslyObserver(instance);
     }
 
     @CheckResult
     @NonNull
-    public static Single<FirebaseUser> signInWithCredential(
+    public static Single<AuthResult> signInWithCredentialAuthResult(
             @NonNull FirebaseAuth instance, @NonNull AuthCredential credential) {
         return new SignInWithCredentialObserver(instance, credential);
     }
 
     @CheckResult
     @NonNull
-    public static Single<FirebaseUser> signInWithCustomToken(
+    public static Single<AuthResult> signInWithCustomTokenAuthResult(
             @NonNull FirebaseAuth instance, @NonNull String token) {
         return new SignInWithCustomTokenObserver(instance, token);
     }
 
     @CheckResult
     @NonNull
-    public static Single<FirebaseUser> signInWithEmailAndPassword(
+    public static Single<AuthResult> signInWithEmailAndPasswordAuthResult(
             @NonNull FirebaseAuth instance, @NonNull String email, @NonNull String password) {
         return new SignInWithEmailAndPasswordObserver(instance, email, password);
+    }
+
+    @CheckResult
+    @NonNull
+    public static Single<FirebaseUser> signInAnonymously(
+            @NonNull FirebaseAuth instance) {
+        return new SignInAnonymouslyObserver(instance)
+                .map(new Function<AuthResult, FirebaseUser>() {
+                    @Override
+                    public FirebaseUser apply(AuthResult authResult) throws Exception {
+                        return authResult.getUser();
+                    }
+                });
+    }
+
+    @CheckResult
+    @NonNull
+    public static Single<FirebaseUser> signInWithCredential(
+            @NonNull FirebaseAuth instance, @NonNull AuthCredential credential) {
+        return new SignInWithCredentialObserver(instance, credential)
+                .map(new Function<AuthResult, FirebaseUser>() {
+                    @Override
+                    public FirebaseUser apply(AuthResult authResult) throws Exception {
+                        return authResult.getUser();
+                    }
+                });
+    }
+
+    @CheckResult
+    @NonNull
+    public static Single<FirebaseUser> signInWithCustomToken(
+            @NonNull FirebaseAuth instance, @NonNull String token) {
+        return new SignInWithCustomTokenObserver(instance, token)
+                .map(new Function<AuthResult, FirebaseUser>() {
+                    @Override
+                    public FirebaseUser apply(AuthResult authResult) throws Exception {
+                        return authResult.getUser();
+                    }
+                });
+    }
+
+    @CheckResult
+    @NonNull
+    public static Single<FirebaseUser> signInWithEmailAndPassword(
+            @NonNull FirebaseAuth instance, @NonNull String email, @NonNull String password) {
+        return new SignInWithEmailAndPasswordObserver(instance, email, password)
+                .map(new Function<AuthResult, FirebaseUser>() {
+                    @Override
+                    public FirebaseUser apply(AuthResult authResult) throws Exception {
+                        return authResult.getUser();
+                    }
+                });
     }
 
     @CheckResult