+++ firebase-auth/src/main/java/com/androidhuman/rxfirebase2/auth/SignInWithCredentialObserver.java
@@ -4,7 +4,6 @@ import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthCredential;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
 
 import com.androidhuman.rxfirebase2.core.OnCompleteDisposable;
 
@@ -13,7 +12,7 @@ import android.support.annotation.NonNull;
 import io.reactivex.Single;
 import io.reactivex.SingleObserver;
 
-final class SignInWithCredentialObserver extends Single<FirebaseUser> {
+final class SignInWithCredentialObserver extends Single<AuthResult> {
 
     private final FirebaseAuth instance;
 
@@ -25,7 +24,7 @@ final class SignInWithCredentialObserver extends Single<FirebaseUser> {
     }
 
     @Override
-    protected void subscribeActual(SingleObserver<? super FirebaseUser> observer) {
+    protected void subscribeActual(SingleObserver<? super AuthResult> observer) {
         Listener listener = new Listener(observer);
         observer.onSubscribe(listener);
 
@@ -33,11 +32,11 @@ final class SignInWithCredentialObserver extends Single<FirebaseUser> {
                 .addOnCompleteListener(listener);
     }
 
-    static final class Listener extends OnCompleteDisposable<AuthResult> {
+    private final class Listener extends OnCompleteDisposable<AuthResult> {
 
-        private final SingleObserver<? super FirebaseUser> observer;
+        private final SingleObserver<? super AuthResult> observer;
 
-        Listener(@NonNull SingleObserver<? super FirebaseUser> observer) {
+        Listener(@NonNull SingleObserver<? super AuthResult> observer) {
             this.observer = observer;
         }
 
@@ -47,9 +46,10 @@ final class SignInWithCredentialObserver extends Single<FirebaseUser> {
                 if (!task.isSuccessful()) {
                     observer.onError(task.getException());
                 } else {
-                    observer.onSuccess(task.getResult().getUser());
+                    observer.onSuccess(task.getResult());
                 }
             }
         }
     }
+
 }