+++ firebase-auth/src/main/java/com/androidhuman/rxfirebase2/auth/SignInWithEmailAndPasswordObserver.java
@@ -3,7 +3,6 @@ package com.androidhuman.rxfirebase2.auth;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.auth.FirebaseUser;
 
 import com.androidhuman.rxfirebase2.core.OnCompleteDisposable;
 
@@ -12,7 +11,7 @@ import android.support.annotation.NonNull;
 import io.reactivex.Single;
 import io.reactivex.SingleObserver;
 
-final class SignInWithEmailAndPasswordObserver extends Single<FirebaseUser> {
+final class SignInWithEmailAndPasswordObserver extends Single<AuthResult> {
 
     private final FirebaseAuth instance;
 
@@ -27,7 +26,7 @@ final class SignInWithEmailAndPasswordObserver extends Single<FirebaseUser> {
     }
 
     @Override
-    protected void subscribeActual(SingleObserver<? super FirebaseUser> observer) {
+    protected void subscribeActual(SingleObserver<? super AuthResult> observer) {
         Listener listener = new Listener(observer);
         observer.onSubscribe(listener);
 
@@ -35,11 +34,11 @@ final class SignInWithEmailAndPasswordObserver extends Single<FirebaseUser> {
                 .addOnCompleteListener(listener);
     }
 
-    static final class Listener extends OnCompleteDisposable<AuthResult> {
+    private final class Listener extends OnCompleteDisposable<AuthResult> {
 
-        private final SingleObserver<? super FirebaseUser> observer;
+        private final SingleObserver<? super AuthResult> observer;
 
-        Listener(@NonNull SingleObserver<? super FirebaseUser> observer) {
+        Listener(@NonNull SingleObserver<? super AuthResult> observer) {
             this.observer = observer;
         }
 
@@ -49,7 +48,7 @@ final class SignInWithEmailAndPasswordObserver extends Single<FirebaseUser> {
                 if (!task.isSuccessful()) {
                     observer.onError(task.getException());
                 } else {
-                    observer.onSuccess(task.getResult().getUser());
+                    observer.onSuccess(task.getResult());
                 }
             }
         }