+++ firebase-auth/src/main/java/com/androidhuman/rxfirebase2/auth/UserReloadObserver.java
@@ -27,7 +27,7 @@ final class UserReloadObserver extends Completable {
                 .addOnCompleteListener(listener);
     }
 
-    static final class Listener extends OnCompleteDisposable<Void> {
+    private static final class Listener extends OnCompleteDisposable<Void> {
 
         private final CompletableObserver observer;