+++ firebase-auth/src/main/java/com/androidhuman/rxfirebase2/auth/UserUpdateEmailObserver.java
@@ -30,7 +30,7 @@ final class UserUpdateEmailObserver extends Completable {
                 .addOnCompleteListener(listener);
     }
 
-    static final class Listener extends OnCompleteDisposable<Void> {
+    private static final class Listener extends OnCompleteDisposable<Void> {
 
         private final CompletableObserver observer;