+++ firebase-auth/src/main/java/com/androidhuman/rxfirebase2/auth/UserUpdateProfileObserver.java
@@ -31,7 +31,7 @@ final class UserUpdateProfileObserver extends Completable {
                 .addOnCompleteListener(listener);
     }
 
-    static final class Listener extends OnCompleteDisposable<Void> {
+    private static final class Listener extends OnCompleteDisposable<Void> {
 
         private final CompletableObserver observer;